US110195 Add assignment instructions methods
These allow fetching and editing of the `instructions` on an assignment entity. There are three different representations of the instructions, all of which are made available:

- `instructionsPlaintext`: self-explanatory, this is the HTML-stripped version of the instructions (not really that useful)
- `instructionsHtml`: HTML version of the instructions for displaying the assignment instructions
- `instructionsEditorHtml`: HTML version of the instructions for usage in editors

An example of the difference between the latter two is that an `{orgUnitId}` string template in the former will show as e.g. "6606", whereas in the latter it will just be "{orgUnitId}".