Add SirenAction ES6 module
This is basically a copy of the existing `siren-fetch-mixin-lit`, with the intent of replacing it entirely. (It will be removed entirely in a subsequent PR.) The intent of this ES6 module is to be able to do "Siren things" anywhere - right now limited to just `performSirenAction`, but potentially with more functionality to come. This can then be dropped into any class that extends `Entity`.

One difference between this and `siren-fetch-mixin-lit` is that `SirenAction.performSirenAction` doesn't emit `d2l-siren-entity-save-*` events. An HTMLElement is required to fire these events, and passing it in was weird and blurred the line of what exactly `siren-sdk` does - so instead, consumers of `SirenAction.performSirenAction` will be responsible for their own start/end/error behaviour, which is easily implemented. This keeps a clearer line between "Siren things" (which `siren-sdk` does) and ancillary behaviour (which is the responsibility of the consumer).