US110198 Add methods for submission type
This simply surfaces the current submission type for assignment, the list of submission type options, and a method to update the current submission type. Note that this doesn't include full support for completion type (coming soon), which means that some options (On Paper and Observed in Person) will always use the default completion type of "Automatically on due date".