US110198 Fixup setSubmissionType
When setting the submissionType to one that doesn't have a completion type, we must set the `completionType` to 0 in the request, otherwise we get a 400 from the LMS (i.e. we always have to set both the submissionType and completionType in the request, "clearing" the latter when it doesn't apply).