US111410 Add support for adding due date
This uses the new `start-add-new` workflow to set an initial due date if one isn't already set. This is exposed via the same `setDueDate` method, meaning the consumer is blind to this change, and the `siren-sdk` will either set or update the due date as required.

This also removes a method that returns a Siren action, as this was added in error - `siren-sdk` should not expose Siren objects.