US111526 Add completion type support
This adds support for completion type values, and slightly modifies how the submission type works - in particular, you should be able to set a submission type without having to also set the completion type - we can silently just set it to the first completion type value as necessary, as a sort of fallback/default.

On the other hand, when setting the completion type, the assumption is that we are _only_ modifying the completion type, i.e. the function is only called when setting a completion type is a valid thing to do.