US111949 Add AttachmentCollectionEntity
This is an entity representation of an attachment collection. Currently, this is only being used in the assignments domain, but the design of this entity is generic enough that it could easily be applied to the activities domain (decisions underway on that). For now, this just gives us the basic abilities we need to show an attachment-picker component, in particular knowing whether or not certain actions are available to the user.