Test the return value of the Add and Delete availability functions
Initially, I attempted to spy/mock the imported `performSirenAction` function from `SirenAction.js`, to test what arguments I was calling it with. The tests passed locally, but not on travis.
I've given up and decided to test the return value of the function instead and skip testing what it calls `performSirenAction` with.