US112361 Add function to attach temp file
Since we're using the file selector dialog in FACE, it takes care of the actual uploading of files. Once they're uploaded by the dialog, we need to attach the existing temp file to the attachment collection, which can be accomplished via the `add-file` action with a `tempFileId` in the POST body.