US111909 Add actions for audio notes
This adds `canAddAudioNoteAttachment` and `addAudioNoteAttachment`, which do as you would expect. This also includes a small refactor to break out some common code, since there are really only two types of attachments:

Files:
- file attachment
- video note attachment
- audio note attachment

Links:
- URL attachment
- Google Drive attachment
- OneDrive attachment

The helpers added here centralize a bit of the logic common to each of these groupings.