Fix bug where a newly associated grade does call `score-out-of`
If a grade is newly associated, ALWAYS call `score-out-of`. The previous behavior did not call `score-out-of` if the previous score and current score are the same.
We need the new behavior because if the activity is now associated to a different grade, we expect the grade to be updated with the activity out of value, which means `score-out-of` must be called when the grade the activity is associated to, changes.

This can happen if the existing activity out of value was say, 50.
We associate to a grade that has a maxPoints of 100.
We change the out of value back to 50 so that when `save` is called the value is 50.
Old behavior would have been to skip the `score-out-of` call. New behavior would be to correctly call `score-out-of` so that the new grade we are associated to gets the value 50.