Renaming quiz subentity specific class to be more generic
The reason for renaming the `hintsEnabled` constant, is that more quiz subentities are going to be added.

I also renamed the `enabled` string to `checked`. I am working on US120983 which adds the subentity "DisableRightClick". If DisableRightClick is checked, it can be confusing to include an `enabled` class, likewise a `disabled` class when it is unchecked (i.e. right click not actually disabled).
The backend change is in this PR: https://github.com/Brightspace/lms/pull/3446