Update working copy checkin and checkout actions
The checkout function will also support fork and the checkin function will also support merge for now. When we update the API to remove fork and merge, I will remove the support from this function.
We are also temporarily supporting a getCheckoutHref function as the checkout action is currently a GET. It will be a POST after we make some changes to the API, at which point, I will remove the getCheckoutHref function.