US123692 - changes to support working copy dirty check
The reaon for changing the recently added `headerHtml` and `descriptionHtml` functions is that the functions return booleans instead of undefined. When a quiz entity which does not have these entities, such as a working copy, calls the `equals` function, the header and description data passed to the `equals` function is `false`, while it compares against `this.descriptionEditorHtml()` and `this.headerEditorHtml()` which will return undefined as the header and description entities do not exist.

Changing `_canCheckin` to not be marked private as it will be called by the `activities` repo.