US123692 - EntityStore should not be updating entity with a different self href.
In working copy, a quiz entity contains a checkout action with a quiz href. After POSTing to it, the response is the working copy entity that ends with a workingCopy queryParam. Wtihout my change, the EntityStore was updating the quiz entity with the returned working copy entity, when in fact, it should be adding a new element to the entity store with the href of the returned working copy.

https://rally1.rallydev.com/#/29180338367d/iterationstatus?detail=%2Fuserstory%2F478606416460%2Ftasks